fix(audit): close C2 — u64 overflow in supermajority checks (DA + bridge + finality + types)
DA certificate, PoHA, bridge verify_supermajority, finality has_supermajority:
all used plain u64 multiplication (x * 3 >= y * 2) which wraps silently
when stake > u64::MAX/3. Cast both sides to u128 before multiplication,
matching the pattern already used in consensus_types::verify_commit (:892).
Also fixes the duplicate stake_quorum_threshold in consensus-types/tendermint.rs
(dead file; fix documents intent for when it is activated).

3 regression tests added (certificate, finality, consensus-types).
941 consensus / 166 DA tests pass.

+++ crates/evaporchain-consensus/src/bridge.rs
@@ -76,7 +76,8 @@ impl ValidatorSetCommitment {
             .map(|(stake, _)| stake)
             .sum();
 
-        signing_stake * 3 > self.total_stake * 2
+        // Audit C2: u128 to prevent wrap when total_stake > u64::MAX/2.
+        (signing_stake as u128) * 3 > (self.total_stake as u128) * 2
     }
 
     /// Verify BLS aggregate signature on a commit certificate against this validator set.
